fix a situation that R doesn't handle well add a fix for AWQMS bug
During the transition from daylight time to standard there is always a repeat of time between 1:00 AM and 1:59 AM. The first pass of time during this period should always be daylight time and the second pass is standard time. Unfortunately when time is forced to UTC (as it is here) R can't tell the first pass from the second and always assumes standard time. This means the first set of timestamps have the wrong UTC offset and needs +1 hour added. This commit fixes this situation.

A new argument was added 'awqms_bug' as a work around to a bug in AWQMS.