Implemnting datetimes
This update categorizes start and end days and times for activity and analtical days and times.

From SH: Activity Start time (equivalent to sample time for us) which is 00:00 of the day of the reported for summary stats
Activity End time which is the last recording of the day of the reported result
Analytical start time which is the min time of the first day  of the considered statistic (different from the rest for MA values)
Analytical end time which is the max time of the last day of the considered statistic, usually the same as the activity end time.